+++ NEWS.md
@@ -89,3 +89,4 @@
 
 # Racmacs 1.1.21
 * Full precision used in rapidjson when parsing doubles to load and save maps
+* Added `numSeraGroups()` function